refactor(http): extract propagator/error helpers from fetch.ts into shared utils
Pulls determinePropagatorTypes, addTraceContextHttpHeaders, endSpanOnError,
and endSpanOnAbort out of fetch.ts into the transport-agnostic
instrumentations/http/utils.ts module so xhr.ts (ENG-4822) can reuse them
without duplicating logic. Pure refactor, no behavior change.

Ref: ENG-4822
